+++ app/src/main/java/app/meeplebook/core/plays/model/Player.kt
@@ -14,10 +14,10 @@ package app.meeplebook.core.plays.model
  * @param win Whether the player won.
  */
 data class Player(
-    val id: Int = 0,
-    val playId: Int,
+    val id: Long = 0,
+    val playId: Long,
     val username: String?,
-    val userId: Int?,
+    val userId: Long?,
     val name: String,
     val startPosition: String?,
     val color: String?,
